requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
langchain
PyPDF2
chromadb
faiss-cpu
langchain_google_genai
langchain-community
python-multipart
